+++ examples/test-resources/main.tf
@@ -11,9 +11,11 @@ terraform {
 }
 
 provider "xenserver" {
-  host     = local.env_vars["XENSERVER_HOST"]
-  username = local.env_vars["XENSERVER_USERNAME"]
-  password = local.env_vars["XENSERVER_PASSWORD"]
+  host         = local.env_vars["XENSERVER_HOST"]
+  username     = local.env_vars["XENSERVER_USERNAME"]
+  password     = local.env_vars["XENSERVER_PASSWORD"]
+  insecure     = local.env_vars["XENSERVER_INSECURE"]
+  ca_cert_path = local.env_vars["XENSERVER_CA_CERT_PATH"]
 }
 
 data "xenserver_sr" "sr" {